pyproject.toml
[project]
name = "cardinal"
version = "0.1.0"
description = "Cardinal"
requires-python = ">=3.12,<3.13"
dependencies = [
    "click>=8.1",
    "PyGithub>=2.1",
    "python-dotenv>=1.0",
]

[project.scripts]
cardinal = "cardinal.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cardinal"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pyright>=1.1",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
typeCheckingMode = "standard"

[tool.pytest.ini_options]
testpaths = ["tests"]
